requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
google-generativeai>=0.3.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.10.0
python-dotenv>=1.0.0
python-multipart>=0.0.5
Pillow>=9.0.0
aiofiles>=23.0.0
pydantic>=2.0.0
requests>=2.28.0
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0
email-validator>=1.0.0
fastapi
pydantic-settings
pyodbc
python-dotenv
sqlalchemy
uvicorn
